# Use reversed_batch option in WAL