--- a/rebar.config
+++ b/rebar.config
@@ -1,5 +1,5 @@
 {deps, [
-        {gen_batch_server, "0.8.0"},
+        {gen_batch_server, "0.8.2"},
         {aten, "0.5.2"}
 ]}.