--- a/src/ra_file_handle.erl
+++ b/src/ra_file_handle.erl
@@ -8,7 +8,7 @@
 -export([init/1, handle_call/3, handle_cast/2, handle_info/2,
          terminate/2, code_change/3]).
 
--export([open/2, close/1, sync/1, write/2, read/2, position/2]).
+-export([open/2, close/1, sync/1, datasync/1, write/2, read/2, position/2]).
 -export([pwrite/2, pwrite/3, pread/2, pread/3]).
 
 -define(SERVER, ?MODULE).
@@ -28,6 +28,9 @@ close(Fd) ->
 
 sync(Fd) ->
     update(io_sync, fun() -> file:sync(Fd) end).
+
+datasync(Fd) ->
+    update(io_sync, fun() -> file:datasync(Fd) end).
 
 write(Fd, Bytes) ->
     update(io_write, iolist_size(Bytes), fun() -> file:write(Fd, Bytes) end).