--- a/src/ra_log_wal.erl
+++ b/src/ra_log_wal.erl
@@ -34,10 +34,18 @@
 % tables and segment notification
 -type writer_id() :: {binary(), pid()}.
 
+-record(batch_writer, {tbl_start :: ra_index(),
+                       uid :: term(),
+                       tid :: term(), %% TODO
+                       from :: ra_index(),
+                       to :: ra_index(),
+                       term :: ra_term(),
+                       inserts = [] :: list()}).
+
 -record(batch, {writes = 0 :: non_neg_integer(),
-                waiting = #{} :: #{pid() =>
-                                   {From :: ra_index(), To :: ra_index(),
-                                    Term :: ra_term()}},
+                waiting = #{} :: #{pid() => #batch_writer{}},
+                                   % {From :: ra_index(), To :: ra_index(),
+                                   %  Term :: ra_term()}},
                 start_time :: maybe(integer()),
                 pending = [] :: iolist()
                }).
@@ -210,7 +218,7 @@ init(#{dir := Dir} = Conf0) ->
 handle_batch(Ops, State0) ->
     State = lists:foldr(fun handle_op/2, start_batch(State0), Ops),
     %% process all ops
-    complete_batch(State).
+    {ok, [garbage_collect], complete_batch(State)}.
 
 terminate(_Reason, State) ->
     _ = cleanup(State),
@@ -372,23 +380,70 @@ handle_msg(rollover, State) ->
     roll_over(State).
 
 append_data(#state{file_size = FileSize,
-                   batch = Batch,
+                   batch = Batch0,
                    writers = Writers} = State,
             {UId, Pid}, Idx, Term, Entry, DataSize, Data, Truncate) ->
-    true = update_mem_table(ra_log_open_mem_tables, UId, Idx, Term, Entry,
-                            Truncate),
+    Batch = incr_batch(ra_log_open_mem_tables, Batch0, UId, Pid,
+                       {Idx, Term}, Data, Entry, Truncate),
     State#state{file_size = FileSize + DataSize,
-                batch = incr_batch(Batch, Pid, {Idx, Term}, Data),
+                batch = Batch,
                 writers = Writers#{UId => {in_seq, Idx}} }.
 
-incr_batch(#batch{writes = Writes,
-                  waiting = Waiting0,
-                  pending = Pend} = Batch, Pid, {Idx, Term}, Data) ->
+incr_batch(OpnMemTbl, #batch{writes = Writes,
+                             waiting = Waiting0,
+                             pending = Pend} = Batch,
+           UId, Pid, {Idx, Term}, Data, Entry, Truncate) ->
     Waiting = case Waiting0 of
-                  #{Pid := {From, _, _}} ->
-                      Waiting0#{Pid => {min(Idx, From), Idx, Term}};
+                  #{Pid := #batch_writer{tbl_start = TblStart0,
+                                         tid = _Tid,
+                                         from = From,
+                                         inserts = Inserts0} = W} ->
+                      % ct:pal("insertingf ~w ~w", [Tid, Idx]),
+                      % true = ets:insert(Tid, {Idx, Term, Entry}),
+                      TblStart = case Truncate of
+                                     true ->
+                                         Idx;
+                                     false ->
+                                         % take the min of the First item in
+                                         % case we are overwriting before
+                                         % the previously first seen entry
+                                         min(TblStart0, Idx)
+                                 end,
+                      Inserts = [{Idx, Term, Entry} | Inserts0],
+                      Waiting0#{Pid => W#batch_writer{from = min(Idx, From),
+                                                      tbl_start = TblStart,
+                                                      to = Idx,
+                                                      term = Term,
+                                                      inserts = Inserts}};
                   _ ->
-                      Waiting0#{Pid => {Idx, Idx, Term}}
+                      %% no batch_writer
+                      {Tid, TblStart} = case ets:lookup(OpnMemTbl, UId) of
+                                            [{_UId, TblStart0, _To, T}] ->
+                                                {T, case Truncate of
+                                                        true ->
+                                                            Idx;
+                                                        false ->
+                                                            min(TblStart0, Idx)
+                                                    end};
+                                            _ ->
+                                                %% there is no table so need
+                                                %% to open one
+                                                T = open_mem_table(UId),
+                                                true = ets:insert_new(
+                                                         OpnMemTbl,
+                                                         {UId, Idx, Idx, T}),
+                                                {T, Idx}
+                                        end,
+                      % ct:pal("inserting ~w ~w", [Tid, Idx]),
+                      % true = ets:insert(Tid, {Idx, Term, Entry}),
+                      Writer = #batch_writer{tbl_start = TblStart,
+                                             from = Idx,
+                                             to = Idx,
+                                             tid = Tid,
+                                             uid = UId,
+                                             term = Term,
+                                             inserts = [{Idx, Term, Entry}]},
+                      Waiting0#{Pid => Writer}
               end,
 
     Batch#batch{writes = Writes + 1,
@@ -426,8 +481,8 @@ update_mem_table(OpnMemTbl, UId, Idx, Term, Entry, Truncate) ->
     end.
 
 roll_over(State0) ->
-    State = flush_pending(State0),
-    roll_over(ra_log_open_mem_tables, State).
+    State = complete_batch(State0),
+    roll_over(ra_log_open_mem_tables, start_batch(State)).
 
 roll_over(OpnMemTbls, #state{wal = Wal0, file_num = Num0,
                              conf = #conf{dir = Dir,
@@ -567,13 +622,21 @@ complete_batch(#state{batch = #batch{waiting = Waiting,
     State = State0#state{metrics_cursor = NextCursor,
                          batch = undefined},
 
-
-    %% notify writers
-    _ = maps:map(fun (Pid, WrittenInfo) ->
-                         Pid ! {ra_log_event, {written, WrittenInfo}},
+    %% process writers
+    _ = maps:map(fun (Pid, #batch_writer{tbl_start = TblStart,
+                                         uid = UId,
+                                         from = From,
+                                         to = To,
+                                         term = Term,
+                                         inserts = Inserts,
+                                         tid = Tid}) ->
+                         true = ets:insert(Tid, Inserts),
+                         true = ets:update_element(ra_log_open_mem_tables, UId,
+                                                   [{2, TblStart}, {3, To}]),
+                         Pid ! {ra_log_event, {written, {From, To, Term}}},
                          ok
                  end, Waiting),
-    {ok, [garbage_collect], State}.
+    State.
 
 wal2list(File) ->
     Data = open_existing(File),