With this option in gen_batch_server 0.8.2 we can avoid reversing the
accumulated batch before processing it using lists:foldr instead. This
results in lower reductions and somewhat faster processing.

Also avoid lists:reverse/1 when building up iodata to be written.

Make sync method configurable

And default to the slightly faster fdatasync instead of fsync after each
batch.
 

Some results using the `ra_log_wal_SUITE:write_many/1` test.

Before:

```
2019-11-01 09:24:31.413
Scenario 10k/8k took 197ms using 432726 reductions for 10000 writes @ 1024 bytes, batch size 8000
Scenario 100k/8k took 1148ms using 4528187 reductions for 100000 writes @ 1024 bytes, batch size 8000
Scenario 200k/4k took 2198ms using 8377454 reductions for 200000 writes @ 1024 bytes, batch size 4000
Scenario 200k/8k took 2147ms using 8992356 reductions for 200000 writes @ 1024 bytes, batch size 8000
Scenario 200k/16k took 2132ms using 9503895 reductions for 200000 writes @ 1024 bytes, batch size 16000
Scenario 200k/32k took 2283ms using 10443741 reductions for 200000 writes @ 1024 bytes, batch size 32000
Scenario 200k/64k took 2368ms using 10951011 reductions for 200000 writes @ 1024 bytes, batch size 64000

2019-11-01 09:25:31.271
Scenario 10k/8k took 234ms using 432835 reductions for 10000 writes @ 1024 bytes, batch size 8000
Scenario 100k/8k took 1129ms using 4586502 reductions for 100000 writes @ 1024 bytes, batch size 8000
Scenario 200k/4k took 2259ms using 8557122 reductions for 200000 writes @ 1024 bytes, batch size 4000
Scenario 200k/8k took 2106ms using 9160543 reductions for 200000 writes @ 1024 bytes, batch size 8000
Scenario 200k/16k took 2176ms using 9606576 reductions for 200000 writes @ 1024 bytes, batch size 16000
Scenario 200k/32k took 2137ms using 10072122 reductions for 200000 writes @ 1024 bytes, batch size 32000
Scenario 200k/64k took 2423ms using 11014812 reductions for 200000 writes @ 1024 bytes, batch size 64000

2019-11-01 09:26:01.059
Scenario 10k/8k took 201ms using 433606 reductions for 10000 writes @ 1024 bytes, batch size 8000
Scenario 100k/8k took 1162ms using 4597763 reductions for 100000 writes @ 1024 bytes, batch size 8000
Scenario 200k/4k took 2283ms using 8558831 reductions for 200000 writes @ 1024 bytes, batch size 4000
Scenario 200k/8k took 2207ms using 9070146 reductions for 200000 writes @ 1024 bytes, batch size 8000
Scenario 200k/16k took 2057ms using 9537743 reductions for 200000 writes @ 1024 bytes, batch size 16000
Scenario 200k/32k took 2283ms using 10304537 reductions for 200000 writes @ 1024 bytes, batch size 32000
Scenario 200k/64k took 2430ms using 10988927 reductions for 200000 writes @ 1024 bytes, batch size 64000
```
After:
```
2019-11-01 09:26:44.034
Scenario 10k/8k took 228ms using 387334 reductions for 10000 writes @ 1024 bytes, batch size 8000
Scenario 100k/8k took 935ms using 3582684 reductions for 100000 writes @ 1024 bytes, batch size 8000
Scenario 200k/4k took 2181ms using 7460133 reductions for 200000 writes @ 1024 bytes, batch size 4000
Scenario 200k/8k took 1790ms using 7209450 reductions for 200000 writes @ 1024 bytes, batch size 8000
Scenario 200k/16k took 1944ms using 8375545 reductions for 200000 writes @ 1024 bytes, batch size 16000
Scenario 200k/32k took 2025ms using 8830944 reductions for 200000 writes @ 1024 bytes, batch size 32000
Scenario 200k/64k took 2152ms using 9355878 reductions for 200000 writes @ 1024 bytes, batch size 64000

2019-11-01 09:27:09.712
Scenario 10k/8k took 206ms using 389273 reductions for 10000 writes @ 1024 bytes, batch size 8000
Scenario 100k/8k took 1010ms using 3607352 reductions for 100000 writes @ 1024 bytes, batch size 8000
Scenario 200k/4k took 2206ms using 7566158 reductions for 200000 writes @ 1024 bytes, batch size 4000
Scenario 200k/8k took 1826ms using 7227331 reductions for 200000 writes @ 1024 bytes, batch size 8000
Scenario 200k/16k took 1966ms using 8366755 reductions for 200000 writes @ 1024 bytes, batch size 16000
Scenario 200k/32k took 2099ms using 8974853 reductions for 200000 writes @ 1024 bytes, batch size 32000
Scenario 200k/64k took 2210ms using 9467563 reductions for 200000 writes @ 1024 bytes, batch size 64000

2019-11-01 09:27:35.841
Scenario 10k/8k took 193ms using 383848 reductions for 10000 writes @ 1024 bytes, batch size 8000
Scenario 100k/8k took 1060ms using 3619378 reductions for 100000 writes @ 1024 bytes, batch size 8000
Scenario 200k/4k took 2243ms using 7665120 reductions for 200000 writes @ 1024 bytes, batch size 4000
Scenario 200k/8k took 1888ms using 7216105 reductions for 200000 writes @ 1024 bytes, batch size 8000
Scenario 200k/16k took 1918ms using 8214699 reductions for 200000 writes @ 1024 bytes, batch size 16000
Scenario 200k/32k took 2078ms using 9082150 reductions for 200000 writes @ 1024 bytes, batch size 32000
Scenario 200k/64k took 2222ms using 9496629 reductions for 200000 writes @ 1024 bytes, batch size 64000
```