@kjnilsson 2019-10-30
dialyzer fails for some reason now. Will investigate tomorrow.