@michaelklishin 2019-11-05
Note: there is one Travis failure due to a timing-sensitive test which is also present in master.